Updated these definitions:
Policy A set of rules defining what is allowed, disallowed or obligatory over an Asset.
Rule A set of actions defining the precise Permissions, Prohibitions and Obligations over an Asset.
Action A set of operations that can be allowed by Permissions, disallowed by Prohibitions, or made obligatory by Duties.
Permission A set of actions that are allowed to be performed over an Asset.
Prohibition A set of actions that are forbidden to be performed over an Asset.
Duty A set of actions that are obliged to be performed over an Asset.
Asset The subject of a Policy that permissions and prohibitions are applied to.
Party Any form of identifiable entity that is the object of a Policy, e.g., a person, a group of people, an organisation, or an agent.
Constraint The limits and restrictions to the Permissions, the Prohibitions and the Duties expressed in the Policy.